==> notebooks/3_SpotifyFeatures.ipynb <==
### Get audio features for the 10,000 songs using the Spotify API. Use the data scraped from Billboard to assign target variable.

import pickle
with open('dataframes/msd_clean.pkl', 'rb') as picklefile:
    msd = pickle.load(picklefile) # load the 10,000 songs

tracks = list(msd.title)

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import time
import sys
import pprint


ids = [] # first get track ids
for i in tracks: 
    if len(i) > 0:
        sp = spotipy.Spotify()
        result = sp.search(q=i, type='track')
        try:
            track_uri = result['tracks']['items'][0]['id']
            ids.append(track_uri)
            time.sleep(.5)
        except:
            ids.append('')
            time.sleep(.5)
    else:
        ids.append('')

import pandas as pd

client_credentials_manager = SpotifyClientCredentials()
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
frames = []
for i in ids: #get audio features from track ids
    audio_features = sp.audio_features(tracks=i)
    df = pd.DataFrame(audio_features)
    frames.append(df)

frames[0:2]

for i, frame in enumerate(frames):
    frame['track_name'] = tracks[i]
    

df = pd.concat(frames)

df.tail(5)

len(df_spotify)

df_spotify['track_title'] = tracks

df_spotify = df_spotify.drop('track_name', axis=1)

#### checks on data 

df_spotify.iloc[6765:6767]

artists = list(msd.artist)

len(artists)

df_tracks['artist'] = artists

df_tracks = df_tracks.drop(df_tracks.columns[-3], axis =1) 

with open('dataframes/billboard_unique.pkl', 'rb') as picklefile:
    bb = pickle.load(picklefile)

### Use Billboard tracks to assign target variable. If track was on Billboard, assign 1, if not, assign 0

bb_tracks = list(bb.track)

def find_target(row):
    lst = bb_tracks
    if row['track_title'] in lst:
        return 1
    else:
        return 0

df_tracks['target'] = df_tracks.apply(find_target, axis=1)

### Looking at class imbalance, we see that 11% of data is target variable

df_tracks.target.value_counts()

with open('dataframes/spotify_tracks_clean.pkl', 'rb') as picklefile:
    df_tracks = pickle.load(picklefile)

df_tracks = df_tracks.convert_objects()

df_tracks.dtypes

import numpy as np
def fillMedian(df):
    return df.fillna(df.median()).dropna(axis=1, how='all')

df_tracks = fillMedian(df_tracks)

df_spotify_model1 = df_tracks.select_dtypes(exclude=[object]).dropna()

df_spotify_model = df_spotify_model1.reset_index(drop=True)

with open('dataframes/spotify_model_data1.pkl', 'wb') as picklefile:
    pickle.dump(df_spotify_model, picklefile)
