==> notebooks/4_add_songs.ipynb <==
### Class imbalance improved by acquiring audio features for 3000 more tracks. 

import pickle
import random
import numpy as np
with open('dataframes/billboard_unique.pkl', 'rb') as picklefile:
    bb = pickle.load(picklefile)

tracks = np.random.choice(bb.track, 3000)

# import spotipy
# from spotipy.oauth2 import SpotifyClientCredentials
# import time
# import sys
# import pprint

# ids = []
# for i in tracks: 
#     if len(i) > 0:
#         sp = spotipy.Spotify()
#         result = sp.search(q=i, type='track')
#         try:
#             track_uri = result['tracks']['items'][0]['id']
#             ids.append(track_uri)
#             time.sleep(.5)
#         except:
#             ids.append('')
#             time.sleep(.5)
#     else:
#         ids.append('')

len(ids)

import pandas as pd

client_credentials_manager = SpotifyClientCredentials()
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
frames = []
for i in ids:
    audio_features = sp.audio_features(tracks=i)
    df = pd.DataFrame(audio_features)
    frames.append(df)

len(frames), len(tracks)

for i, frame in enumerate(frames):
    frame['track_name'] = tracks[i]
    

def find_spotify(row):
    lst = tracks
    if row['track'] in lst:
        return 1
    else:
        return 0
    
bb['target'] = bb.apply(find_spotify, axis=1)

bb_spotify = bb[bb['target']== 1]

bb_spotify = bb_spotify.sort_values(by='track')

with open('dataframes/spotify_df7.pkl', 'rb') as picklefile:
    df = pickle.load(picklefile)

df = df.drop_duplicates(['track_name'], keep='last')

# bb_artists = list(bb_spotify.artist)
# bb_artists

df = df.sort_values(by='track_name')

df['artist'] = bb_artists

df_spotify2 = df.drop(df.columns[-2], axis =1) 

df_spotify3 = df_spotify2.dropna()

df_spotify3 = df_spotify3.convert_objects()

df_spotify4 = df_spotify3.select_dtypes(exclude=[object]).dropna()

#df_spotify4['target'] = 1


with open('dataframes/spotify_model_addon.pkl', 'wb') as picklefile:
    pickle.dump(df_spotify4, picklefile)

with open('dataframes/spotify_model_data1.pkl', 'rb') as picklefile:
    df_tracks = pickle.load(picklefile)

df_tracks.head()

df_master = df_tracks.append(df_spotify4).reset_index(drop=True)

df_master.tail()

with open('dataframes/df_master.pkl', 'wb') as picklefile:
    pickle.dump(df_master, picklefile)
